Make learna installable via pip
- Add setup.py: Learna is installable via ``pip install .`` with this changes. If learna is installed properly via pip, it can be used much easier in different code bases. Otherwise, (currently), you have to append the source directory to the PATH-variable.
- Rename the `src` directory to `learna`.
- Change relative imports to absolute import statements
