+++ utils/execution_scripts/LEARNA-10min.sh
@@ -2,7 +2,7 @@
 TARGET_STRUCTURE_PATH=$1
 
 source thirdparty/miniconda/miniconda/bin/activate learna
-/usr/bin/time -f"%U" python -m src.learna.design_rna \
+/usr/bin/time -f"%U" python -m learna.learna.design_rna \
   --batch_size 32 \
   --conv_channels 8 1 \
   --embedding_size 0 \
